fix comment

+++ dbms/src/Flash/Coprocessor/CoprocessorReader.h
@@ -139,7 +139,8 @@ public:
         return detail;
     }
 
-    CoprocessorReaderResult nextResult(std::queue<Block> & block_queue, const Block & header, UInt32 /* stream_id */)
+    // stream_id is only meaningful for ExchagneReceive.
+    CoprocessorReaderResult nextResult(std::queue<Block> & block_queue, const Block & header, size_t/*stream_id*/)
     {
         auto && [result, has_next] = resp_iter.next();
         if (!result.error.empty())
